+++ drivers/net/ipa/ipa_i.h
@@ -149,7 +149,7 @@ struct ipa_ep_context {
 	bool allocated;
 	enum ipa_client_type client;
 	unsigned long gsi_chan_hdl;
-	unsigned long gsi_evt_ring_hdl;
+	u32 gsi_evt_ring_hdl;
 	bool bytes_xfered_valid;
 	u16 bytes_xfered;
